requirements.txt
streamlit==1.32.0
python-Levenshtein==0.23.0 